requirements.txt
black==22.3.0
click==8.1.2
flake8==4.0.1
Flask==2.1.1
gunicorn==20.1.0
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.1.1
zipp==3.8.0
git+https://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@github.com/EdwinZurawik/flatty@v0.0.4#egg=flatty
